syncscroll: completed syncscroll implementation and moved it to separate file
This implementation tries to keep the line that the cursor is on and the
line in the preview that it corresponds with, next to each other.
If the cursor is outside the viewport it will do this for the line that
is closest to the cursor, but still inside the viewport.

In tab.py preview creation related code was moved to a separate class.