Always use the basename as markup fallback title
If we do not specify baseName=True, the getDocumentTitle function will
use the markup to retrieve the title, which will usually result in the
document being processed twice by the markup.

Also, pass the fallback title always (it does not make much sense to hide
it for the HTML preview).