Simplify the interface between RetextWindow and RetextTab
The fileName member of the tab is now read-only and automatically
updated as part of the save and load functions. This makes it impossible
for the window to forget updating the filename after loading/saving.
Also some signals were added to the tab for file name changes,
modification state changes and active markup changes. Using these
signals the window no longer has to know which of the tab functions it
calls should be accompanied by updates of its own state.