converterprocess: handle the BrokenPipeError we may get while writing
If the application exits while the converter process is busy, the
converter process will get a BrokenPipeError when it tries to write back
the result. The parent process may have queued a 'quit' command before
closing the pipe, so it's still possible that nothing is wrong.
The converter process should ignore the broken pipe on write and try to
read the next command. If a 'quit' command is read then it just exits
and if an EOFerror is raised it should exit with a traceback to show
that something went wrong.