window: Remove aliases from list of available codecs, and sort that list
Sadly Qt provides no API to get the list of available codec names without
aliases (availableMibs is not a solution because some codecs have zero MIB).

So we have to use this hack to get the list of unique codec names.

It significantly reduces the list (from 805 items to 104 for me).