ReTextPreview: fix undefined name error, and better links support
ReTextPreview has no attribute _fileName, so we should not use this.

Also, if a link points to a local .html file, and a matching source file
was found, open that file in a new tab.

Split the logic into ReTextTab.openSourceFile method, so that other
previewers can make use of it.