Add text statistics in the lower left corner of the editor
This adds a word count, alpha numeric character count and all character
count in the bottom left corner

This option can be disabled from the Preferences dialog (by default it's
off). The statistic is only updated when the content changes (to avoid
recomputing when scrolling or resizing the window). On my lapop, the
computing only takes 40ms for a 100'000 characters markdown document
(roughly 15'000 words, 3'500 lines).

Fixes #164