Change the way fonts are stored in settings
Store the string returned by QFont.toString() method, which is a comma-
separated list of attributes, and includes weight, style and other
properties.

This way it is possible to make the editor font bold, and store this
configuration in settings.

Also refactor preview classes to retrieve font from settings only once.

Fixes #555.