Use Python I/O functions instead of QTextStream
We used QTextStream.setCodec(), which is removed in Qt 6. So use the
Python way for reading files and writing to them, and use Python's
encodings instead of QTextCodec.

Unfortunately there is no way to get a list of supported encodings, so
use a line edit with validation instead of a list of items.