Check if auto-save is enabled before saving the files
If a document has been modified externally and the user chose to not
reload the file, we disable auto-save for the session by setting
autoSaveEnabled = False. However that did not really work because the
saveAll function, which is called whenever the timer fires, did not
check the value of autoSaveEnabled attribute.

Fix this by calling autoSaveActive method, which takes autoSaveEnabled
attribute into account.