Move some of the config options to a separate cache file
Some of the settings described in the configuration.md file are said to be editable by
ReText only, not by the user. Therefore, it makes sense to store them in separate files.

One possible scenario would be if the user wanted to save their configuration. Now they can
save the ReText.conf file and use it on another instance of ReText without affecting the
config entries that should not be user editable.