pyproject.toml: Depend on Markups[markdown,restructuredtext]
Markdown markup also needs python-markdown-math module to be installed.

As that is an implementation detail of pymarkups and may change in the
future, use extras syntax to make sure we are always up to date.

Also, bump required Markups version to 3.1, as the extra names were
changed (lower-cased) in that version.